# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AquaForge API",
    "description": "Swim Meet Optimization API - Backend services for AquaForge.ai",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/optimize": {
      "post": {
        "tags": [
          "Optimization (Legacy)"
        ],
        "summary": "Optimize Lineup",
        "description": "Run lineup optimization for a swim meet.\n\nThis endpoint accepts team data and optimization parameters,\nthen returns the optimal lineup assignments.\n\nArgs:\n    request: Optimization request with team data and parameters\n\nReturns:\n    Optimization results with scores and lineup assignments",
        "operationId": "optimize_lineup_api_v1_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/championship/project": {
      "post": {
        "tags": [
          "Championship",
          "Championship"
        ],
        "summary": "Project championship standings",
        "description": "Project meet standings from psych sheet data.\n    \n    Uses seed times to predict placements and calculate expected points\n    for each team. Also identifies swing events where small improvements\n    would yield significant point gains.\n    \n    **Supported Meet Profiles:**\n    - `vcac_championship`: VCAC Conference Championship rules\n    - `visaa_state`: VISAA State Championship rules",
        "operationId": "project_standings_api_v2_championship_project_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChampionshipProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionshipProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/championship/optimize": {
      "post": {
        "tags": [
          "Championship",
          "Championship"
        ],
        "summary": "Optimize championship entries",
        "description": "Optimize swimmer event assignments for maximum team points.\n    \n    Takes into account:\n    - Max 2 individual events per swimmer\n    - Diving counts as 1 individual event\n    - 400 Free Relay (Relay 3) counts as 1 individual event at VCAC\n    - Back-to-back event restrictions",
        "operationId": "optimize_entries_api_v2_championship_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChampionshipOptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionshipOptimizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/championship/upload-psych-sheet": {
      "post": {
        "tags": [
          "Championship",
          "Championship"
        ],
        "summary": "Upload psych sheet file",
        "description": "Upload a CSV psych sheet and parse it.",
        "operationId": "upload_psych_sheet_api_v2_championship_upload_psych_sheet_post",
        "parameters": [
          {
            "name": "target_team",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Seton",
              "title": "Target Team"
            }
          },
          {
            "name": "meet_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Championship",
              "title": "Meet Name"
            }
          },
          {
            "name": "meet_profile",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "vcac_championship",
              "title": "Meet Profile"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_psych_sheet_api_v2_championship_upload_psych_sheet_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Upload Psych Sheet Api V2 Championship Upload Psych Sheet Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/championship/scoring-info/{profile}": {
      "get": {
        "tags": [
          "Championship",
          "Championship"
        ],
        "summary": "Get scoring information for a meet profile",
        "description": "Get scoring information for a specific meet profile.",
        "operationId": "get_scoring_info_api_v2_championship_scoring_info__profile__get",
        "parameters": [
          {
            "name": "profile",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Scoring Info Api V2 Championship Scoring Info  Profile  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_psych_sheet_api_v2_championship_upload_psych_sheet_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_psych_sheet_api_v2_championship_upload_psych_sheet_post"
      },
...
      "ChampionshipOptimizeRequest": {
        "properties": {
          "entries": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Entries",
            "description": "All psych sheet entries"
          },
          "target_team": {
            "type": "string",
            "title": "Target Team",
            "description": "Team to optimize",
            "default": "Seton"
          },
          "meet_name": {
            "type": "string",
            "title": "Meet Name",
            "description": "Meet name",
            "default": "Championship"
          },
          "meet_profile": {
            "type": "string",
            "title": "Meet Profile",
            "description": "Rules profile",
            "default": "vcac_championship"
          },
          "divers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Divers",
            "description": "Swimmers who are diving"
          },
          "relay_3_swimmers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Relay 3 Swimmers",
            "description": "Swimmers swimming the 400 Free Relay"
          },
          "optimize_entries": {
            "type": "boolean",
            "title": "Optimize Entries",
            "description": "Run entry optimization",
            "default": true
          },
          "optimize_relays": {
            "type": "boolean",
            "title": "Optimize Relays",
            "description": "Run relay optimization",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "entries"
        ],
        "title": "ChampionshipOptimizeRequest",
        "description": "Request for championship entry optimization."
      },
...
      "ChampionshipProjectRequest": {
        "properties": {
          "entries": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Entries",
            "description": "All psych sheet entries"
          },
          "target_team": {
            "type": "string",
            "title": "Target Team",
            "description": "Team to focus analysis on",
            "default": "Seton"
          },
          "meet_name": {
            "type": "string",
            "title": "Meet Name",
            "description": "Meet name",
            "default": "Championship"
          },
          "meet_profile": {
            "type": "string",
            "title": "Meet Profile",
            "description": "Rules profile",
            "default": "vcac_championship"
          }
        },
        "type": "object",
        "required": [
          "entries"
        ],
        "title": "ChampionshipProjectRequest",
        "description": "Request for championship point projection.",
        "example": {
          "entries": [
            {
              "event": "100 Free",
              "grade": 11,
              "seed_time": 52.34,
              "swimmer_name": "John Smith",
              "team": "Seton"
            }
          ],
          "meet_name": "VCAC Championship 2026",
          "target_team": "Seton"
        }
      },
      "ChampionshipProjectResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether projection succeeded"
          },
          "meet_name": {
            "type": "string",
            "title": "Meet Name",
            "description": "Meet name"
          },
          "target_team": {
            "type": "string",
            "title": "Target Team",
            "description": "Target team"
          },
          "target_team_total": {
            "type": "number",
            "title": "Target Team Total",
            "description": "Target team projected points"
          },
          "target_team_rank": {
            "type": "integer",
            "title": "Target Team Rank",
            "description": "Target team projected rank"
          },
          "standings": {
            "items": {
              "$ref": "#/components/schemas/StandingModel"
            },
            "type": "array",
            "title": "Standings",
            "description": "Projected standings"
          },
          "team_totals": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Team Totals",
            "description": "Points per team"
          },
          "swing_events": {
            "items": {
              "$ref": "#/components/schemas/SwingEventModel"
            },
            "type": "array",
            "title": "Swing Events",
            "description": "Swing event opportunities"
          },
          "event_projections": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/EventProjectionModel"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Projections",
            "description": "Per-event projections"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "Coaching recommendations"
          },
          "meet_type": {
            "type": "string",
            "title": "Meet Type",
            "description": "Meet type",
            "default": "championship"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics",
            "description": "Performance metrics"
          },
          "warnings": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings",
            "description": "Validation warnings"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if failed"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meet_name",
          "target_team",
          "target_team_total",
          "target_team_rank",
          "standings",
          "team_totals"
        ],
        "title": "ChampionshipProjectResponse",
        "description": "Response from championship point projection."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OptimizationRequest": {
        "properties": {
          "seton_data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Seton Data",
            "description": "Seton team data"
          },
          "opponent_data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Opponent Data",
            "description": "Opponent team data"
          },
          "optimizer_backend": {
            "$ref": "#/components/schemas/OptimizerBackend",
            "description": "Optimization engine to use (GUROBI = accurate MILP, HEURISTIC = fast fallback)",
            "default": "gurobi"
          },
          "max_individual_events": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Max Individual Events",
            "default": 4
          },
          "enforce_fatigue": {
            "type": "boolean",
            "title": "Enforce Fatigue",
            "default": true
          },
          "use_relay_optimization": {
            "type": "boolean",
            "title": "Use Relay Optimization",
            "default": true
          },
          "robust_mode": {
            "type": "boolean",
            "title": "Robust Mode",
            "description": "Run multi-scenario evaluation",
            "default": false
          },
          "scoring_type": {
            "type": "string",
            "title": "Scoring Type",
            "description": "Scoring rules (visaa_top7 or standard_top5)",
            "default": "visaa_top7"
          },
          "seton_team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seton Team Name",
            "default": "Seton"
          },
          "opponent_team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Team Name",
            "default": "Opponent"
          }
        },
        "type": "object",
        "required": [
          "seton_data",
          "opponent_data"
        ],
        "title": "OptimizationRequest",
        "description": "Request model for optimization."
      },
      "OptimizationResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "seton_score": {
            "type": "number",
            "title": "Seton Score"
          },
          "opponent_score": {
            "type": "number",
            "title": "Opponent Score"
          },
          "score_margin": {
            "type": "number",
            "title": "Score Margin"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/OptimizationResult"
            },
            "type": "array",
            "title": "Results"
          },
          "statistics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Statistics"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "optimization_time_ms": {
            "type": "number",
            "title": "Optimization Time Ms"
          },
          "championship_standings": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Championship Standings",
            "description": "Full team standings for championship mode: [{rank, team, points}, ...]"
          },
          "event_breakdowns": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Breakdowns",
            "description": "Per-event point breakdown for all teams"
          },
          "swing_events": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swing Events",
            "description": "Events where small improvements yield significant gains"
          },
          "analytics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analytics",
            "description": "Advanced analytics: monte_carlo, fatigue_warnings, nash_insights, relay_tradeoffs"
          }
        },
        "type": "object",
        "required": [
          "success",
          "seton_score",
          "opponent_score",
          "score_margin",
          "results",
          "optimization_time_ms"
        ],
        "title": "OptimizationResponse",
        "description": "Response from optimization request."
      },
      "OptimizationResult": {
        "properties": {
          "event": {
            "type": "string",
            "title": "Event"
          },
          "event_number": {
            "type": "integer",
            "title": "Event Number"
          },
          "seton_swimmers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Seton Swimmers"
          },
          "opponent_swimmers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Opponent Swimmers"
          },
          "seton_times": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Seton Times"
          },
          "opponent_times": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Opponent Times"
          },
          "seton_points": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Seton Points"
          },
          "opponent_points": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Opponent Points"
          },
          "projected_score": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Projected Score"
          }
        },
        "type": "object",
        "required": [
          "event",
          "event_number",
          "seton_swimmers",
          "opponent_swimmers",
          "seton_times",
          "opponent_times",
          "projected_score"
        ],
        "title": "OptimizationResult",
        "description": "Single event optimization result."
      },
      "OptimizerBackend": {
        "type": "string",
        "enum": [
          "heuristic",
          "gurobi"
        ],
        "title": "OptimizerBackend",
        "description": "Available optimization backends."
      },
...
    }
  }
}